requirements.txt
pytorch==1.1.0
torchvision==0.3.0
numpy==1.14.3
scipy==1.1.0
scikit-learn==0.19.1
tensorboardX